split roadCount from roads submodule
