+++ app/assets/scripts/redux/modules/roads.js
@@ -7,6 +7,9 @@ import {
   property,
   omit
 } from 'lodash';
+import {
+  clearRoadCount
+} from './roadCount';
 import config from '../../config';
 
 
@@ -20,8 +23,6 @@ export const roadIdIsValid = (id, province, district) => {
 };
 export const getRoadPageKey = (province = '', district = '', page, sortField, sortOrder) =>
   `${province}-${district}-${page}-${sortField}-${sortOrder}`;
-export const getRoadCountKey = (province = '', district = '') =>
-  `${province}-${district}`;
 
 
 /**
@@ -31,10 +32,6 @@ export const FETCH_ROADS = 'FETCH_ROADS';
 export const FETCH_ROADS_SUCCESS = 'FETCH_ROADS_SUCCESS';
 export const FETCH_ROADS_ERROR = 'FETCH_ROADS_ERROR';
 export const CLEAR_ROADS_PAGES = 'CLEAR_ROADS_PAGES';
-export const FETCH_ROAD_COUNT = 'FETCH_ROAD_COUNT';
-export const FETCH_ROAD_COUNT_SUCCESS = 'FETCH_ROAD_COUNT_SUCCESS';
-export const FETCH_ROAD_COUNT_ERROR = 'FETCH_ROAD_COUNT_ERROR';
-export const CLEAR_ROAD_COUNT = 'CLEAR_ROAD_COUNT';
 export const FETCH_ROAD_GEOMETRY = 'FETCH_ROAD_GEOMETRY';
 export const FETCH_ROAD_GEOMETRY_SUCCESS = 'FETCH_ROAD_GEOMETRY_SUCCESS';
 export const FETCH_ROAD_GEOMETRY_ERROR = 'FETCH_ROAD_GEOMETRY_ERROR';
@@ -59,13 +56,6 @@ export const fetchRoadsSuccess = (roadsById, roadsByPage, province, district, pa
 export const fetchRoadsError = (error, province, district, page, sortField, sortOrder) =>
   ({ type: FETCH_ROADS_ERROR, error, province, district, page, sortField, sortOrder });
 export const clearRoadsPages = () => ({ type: CLEAR_ROADS_PAGES });
-
-export const fetchRoadCount = (province, district) => ({ type: FETCH_ROAD_COUNT, province, district });
-export const fetchRoadCountSuccess = (count, pageCount, osmCount, province, district) =>
-  ({ type: FETCH_ROAD_COUNT_SUCCESS, count, pageCount, osmCount, province, district });
-export const fetchRoadCountError = (error, province, district) =>
-  ({ type: FETCH_ROAD_COUNT_ERROR, error, province, district });
-export const clearRoadCount = () => ({ type: CLEAR_ROAD_COUNT });
 
 export const fetchRoadGeometry = (id) => ({ type: FETCH_ROAD_GEOMETRY, id });
 export const fetchRoadGeometrySuccess = (id, geoJSON) => ({ type: FETCH_ROAD_GEOMETRY, id, geoJSON });
@@ -108,26 +98,6 @@ export const fetchRoadsEpic = (province, district, page, sortField, sortOrder) =
       dispatch(fetchRoadsSuccess(roadsById, roadsByPage, province, district, page, sortField, sortOrder));
     })
     .catch((err) => dispatch(fetchRoadsError(err, province, district, page, sortField, sortOrder)));
-};
-
-
-export const fetchRoadCountEpic = (province, district) => (dispatch) => {
-  dispatch(fetchRoadCount(province, district));
-
-  return fetch(
-    format({ pathname: `${config.api}/properties/roads/count`, query: { province, district } })
-  )
-    .then(response => {
-      if (!response.ok) {
-        throw new Error(response.status);
-      }
-
-      return response.json();
-    })
-    .then(({ count, pageCount, osmCount }) => {
-      dispatch(fetchRoadCountSuccess(count, pageCount, osmCount, province, district));
-    })
-    .catch((err) => dispatch(fetchRoadCountError(err, province, district)));
 };
 
 
@@ -217,8 +187,7 @@ export const deleteRoadEpic = (id) => (dispatch) => {
 export default (
   state = {
     roadsById: {},
-    roadsByPage: {},
-    roadCount: {}
+    roadsByPage: {}
   },
   action
 ) => {
@@ -262,48 +231,6 @@ export default (
   } else if (action.type === CLEAR_ROADS_PAGES) {
     return Object.assign({}, state, {
       roadsByPage: {}
-    });
-  } else if (action.type === FETCH_ROAD_COUNT) {
-    const { province, district } = action;
-    const roadCountKey = getRoadCountKey(province, district);
-
-    return Object.assign({}, state, {
-      roadCount: Object.assign({}, state.roadCount, {
-        [roadCountKey]: Object.assign({}, state.roadCount[roadCountKey] || {}, {
-          status: 'pending'
-        })
-      })
-    });
-  } else if (action.type === FETCH_ROAD_COUNT_SUCCESS) {
-    const { province, district, count, pageCount, osmCount } = action;
-    const roadCountKey = getRoadCountKey(province, district);
-
-    return Object.assign({}, state, {
-      roadCount: Object.assign({}, state.roadCount, {
-        [roadCountKey]: {
-          status: 'complete',
-          count,
-          pageCount,
-          osmCount
-        }
-      })
-    });
-  } else if (action.type === FETCH_ROAD_COUNT_ERROR) {
-    const { province, district } = action;
-    const roadCountKey = getRoadCountKey(province, district);
-
-    return Object.assign({}, state, {
-      roadCount: Object.assign({}, state.roadCount, {
-        [roadCountKey]: {
-          status: 'error',
-          count: undefined,
-          pageCount: undefined
-        }
-      })
-    });
-  } else if (action.type === CLEAR_ROAD_COUNT) {
-    return Object.assign({}, state, {
-      roadCount: {}
     });
   }
